Rename: `commands.rs` => `cli.rs`
